NO AUTO Strips load, bitwise, warmup, stats, data.url, log, label, example from full
