+++ core/src/main/java/apoc/bitwise/BitwiseOperations.java
