+++ core/src/main/java/apoc/load/LoadJson.java
@@ -4,7 +4,6 @@ import apoc.result.MapResult;
 import apoc.result.ObjectResult;
 import apoc.util.CompressionAlgo;
 import apoc.util.JsonUtil;
-import apoc.util.Util;
 import org.neo4j.graphdb.GraphDatabaseService;
 import org.neo4j.procedure.Context;
 import org.neo4j.procedure.Description;
@@ -12,12 +11,11 @@ import org.neo4j.procedure.Name;
 import org.neo4j.procedure.Procedure;
 
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Stream;
 
-import static apoc.util.CompressionConfig.COMPRESSION;
+import static apoc.load.LoadJsonUtils.loadJsonStream;import static apoc.util.CompressionConfig.COMPRESSION;
 
 public class LoadJson {
 
@@ -58,32 +56,6 @@ public class LoadJson {
         String compressionAlgo = (String) config.getOrDefault(COMPRESSION, CompressionAlgo.NONE.name());
         List<String> pathOptions = (List<String>) config.get("pathOptions");
         return loadJsonStream(urlOrKeyOrBinary, headers, payload, path, failOnError, compressionAlgo, pathOptions);
-    }
-
-    public static Stream<MapResult> loadJsonStream(@Name("url") Object url, @Name("headers") Map<String, Object> headers, @Name("payload") String payload) {
-        return loadJsonStream(url, headers, payload, "", true, null, null);
-    }
-    public static Stream<MapResult> loadJsonStream(@Name("urlOrKeyOrBinary") Object urlOrKeyOrBinary, @Name("headers") Map<String, Object> headers, @Name("payload") String payload, String path, boolean failOnError, String compressionAlgo, List<String> pathOptions) {
-        if (urlOrKeyOrBinary instanceof String) {
-            headers = null != headers ? headers : new HashMap<>();
-            headers.putAll(Util.extractCredentialsIfNeeded((String) urlOrKeyOrBinary, failOnError));
-        }
-        Stream<Object> stream = JsonUtil.loadJson(urlOrKeyOrBinary,headers,payload, path, failOnError, compressionAlgo, pathOptions);
-        return stream.flatMap((value) -> {
-            if (value instanceof Map) {
-                return Stream.of(new MapResult((Map) value));
-            }
-            if (value instanceof List) {
-                if (((List)value).isEmpty()) return Stream.empty();
-                if (((List) value).get(0) instanceof Map)
-                    return ((List) value).stream().map((v) -> new MapResult((Map) v));
-                return Stream.of(new MapResult(Collections.singletonMap("result",value)));
-            }
-            if(!failOnError)
-                throw new RuntimeException("Incompatible Type " + (value == null ? "null" : value.getClass()));
-            else
-                return Stream.of(new MapResult(Collections.emptyMap()));
-        });
     }
 
 }
