+++ core/src/main/java/apoc/load/SkipWhitespaceInputStream.java
