+++ core/src/main/java/apoc/log/Neo4jLogStream.java
