+++ core/src/main/java/apoc/stats/DegreeDistribution.java
