+++ core/src/main/java/apoc/stats/DegreeUtil.java
