+++ core/src/test/java/apoc/bitwise/BitwiseOperationsTest.java
