+++ core/src/test/java/apoc/load/SimpleXmlTest.java
