+++ core/src/test/java/apoc/load/XmlTest.java
