+++ core/src/test/java/apoc/log/Neo4jLogStreamTest.java
