+++ core/src/test/java/apoc/scoring/ScoringTest.java
