+++ core/src/test/java/apoc/stats/DegreeDistributionTest.java
