+++ core/src/test/resources/map.json
