+++ core/src/test/resources/xml/books.xml
