+++ core/src/test/resources/xml/booksResult.txt
