+++ core/src/test/resources/xml/databases.xml
