+++ core/src/test/resources/xml/humboldt_soemmering01_1791.TEI-P5-shortened.xml
