+++ core/src/test/resources/xml/missingExternalDTD.xml
