+++ core/src/test/resources/xml/mixedcontent.xml
