+++ core/src/test/resources/xml/singleLine.xml
