+++ core/src/test/resources/xml/xxe.xml
